+++ crates/nbcl_impl/src/lib.rs
@@ -1,0 +1,6 @@
+//! nbcl_impl is a simple crate which configures nbcl for the `ewwii` widget system.
+
+pub mod builtins;
+pub mod errors;
+pub mod parser;
+pub mod translate;